on start load page 1 (20 movies), then loading next pages at scrolling on the end list, fixed small bugs, refreshing only loaded pages number.

+++ app/src/main/java/com/jeksonshar/funacademyapp/network/models/MovieDetailsModel.kt
@@ -18,7 +18,7 @@ data class MovieDetailsModel(
 	@SerialName("backdrop_path")
 	val backdropPath: String? = null,
 	@SerialName("revenue")
-	val revenue: Int? = null,
+	val revenue: Float? = null,
 	@SerialName("genres")
 	val genres: List<Genre>? = null,
 	@SerialName("popularity")
